+++ HISTORY.md
@@ -12,6 +12,7 @@
 * lists: fixed a RESHAPE edge case when passing in a single zero dimension
 * lists: made sure ITEM OF returns data matching the shape specified by the query struct
 * lists: hyperized ITEM OF
+* threads: refactored reportRank()
 
 ## 9.1.1:
 * **Notable Fixes:**